Refactored EntryMgrAction into EntryViewerAction and EntryEditorAction (see ticket #144 for details; updated menu structure, fixed several styling isues

+++ ic-persist/src/main/java/edu/ucla/mbi/imex/central/dao/IcPubDao.java
@@ -611,7 +611,7 @@ public class IcPubDao extends AbstractDAO implements PublicationDao {
             }
             
             List foo = crit.setProjection( Projections.rowCount()).list();
-            count  = ((Integer) foo.get(0) ).longValue() ;
+            count  = ((Long) foo.get(0) ).longValue() ;
 
             Log log = LogFactory.getLog( this.getClass() );
             log.debug( "count=" + count );
